fix: 유저 api 수정
게스트 계정 생성 시 user Role로 저장되던 것 수정.
id 타입 integer로 통일.
/user/delete 서비스 로직 예외처리 추가.

+++ src/main/java/com/springboot/todo/controller/UserController.java
@@ -47,7 +47,7 @@ public class UserController {
 
 
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<HttpStatus> deleteUser(@PathVariable("id") Long id) {
+    public ResponseEntity<HttpStatus> deleteUser(@PathVariable("id") Integer id) {
         try {
             userService.deleteUser(id);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
